Auto-detect the nonce value from the server response
Previously it was impossible to make more than one API request per second
because the client API code used the current second as the nonce value to
pass on each Trade API request.

This change allows the API to run at full speed by incrementing the nonce
on each request. It also establishes the correct nonce value to start
with by interrogating the server during initialisation